# 01 - Pilar 1/orders-api-swagger.yaml
openapi: 3.0.3
info:
  title: Orders Management API
  description: |
    RESTful API for managing customer orders - Workshop Demo
    
    This API provides endpoints to manage customer orders in a simple e-commerce or sales system.
    
    ## Features
    - List all orders
    - Find specific order by ID
    - Standardized response format
    - Error handling
    
    ## Architecture
    Built using Clean Architecture principles with:
    - Domain Layer (entities and business rules)
    - Application Layer (use cases)
    - Infrastructure Layer (persistence)
    - Interface Layer (REST controllers)
  version: 1.0.0
  contact:
    name: BrasilPrev Team
    email: dev@brasilprev.com.br
    url: https://brasilprev.com.br
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8081
    description: Development Server
  - url: https://api.brasilprev.com.br
    description: Production Server

tags:
  - name: Orders
    description: Order management operations

paths:
  /api/v1/orders:
    get:
      tags:
        - Orders
      summary: List all orders
      description: Retrieve a list of all orders in the system
      operationId: listOrders
      responses:
        '200':
          description: Successfully retrieved list of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
              examples:
                empty_list:
                  summary: Empty order list
                  value:
                    success: true
                    data: []
                    message: null
                with_orders:
                  summary: List with orders
                  value:
                    success: true
                    data:
                      - id: 1
                        customerName: "John Silva"
                        status: "PENDING"
                        totalAmount: 150.00
                        createdAt: "2026-03-15T03:01:17"
                      - id: 2
                        customerName: "Mary Souza"
                        status: "APPROVED"
                        totalAmount: 320.50
                        createdAt: "2026-03-15T03:02:30"
                    message: null
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Find order by ID
      description: Retrieve a specific order by its unique identifier
      operationId: findOrderById
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Successfully retrieved the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                success: true
                data:
                  id: 1
                  customerName: "John Silva"
                  status: "PENDING"
                  totalAmount: 150.00
                  createdAt: "2026-03-15T03:01:17"
                message: null
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                data: null
                message: "Order not found: 1"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Order:
      type: object
      description: Order entity representing a customer order
      required:
        - customerName
        - status
        - totalAmount
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the order
          example: 1
        customerName:
          type: string
          description: Name of the customer who placed the order
          example: "John Silva"
          maxLength: 255
        status:
          type: string
          description: Current status of the order
          example: "PENDING"
          enum:
            - PENDING
            - APPROVED
            - CANCELLED
        totalAmount:
          type: number
          format: decimal
          description: Total amount of the order
          example: 150.00
          minimum: 0
        createdAt:
          type: string
          format: date-time
          description: Date and time when the order was created
          example: "2026-03-15T03:01:17"

    ApiResponse:
      type: object
      description: Standard response wrapper for all API endpoints
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        data:
          description: The actual data returned by the operation
        message:
          type: string
          description: Error message if operation failed
          example: "Order not found"
          nullable: true

    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Order'

    OrderListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Order'

    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              example: false
            data:
              example: null
            message:
              type: string
              example: "Order not found: 1"